feat: add PDV, sales report, navbar and product management

+++ src/main/java/com/kaua8/mercadoFacil/service/ProductService.java
@@ -2,7 +2,8 @@ package com.kaua8.mercadoFacil.service;
 
 import com.kaua8.mercadoFacil.model.Product;
 import com.kaua8.mercadoFacil.repository.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.kaua8.mercadoFacil.repository.SaleItemRepository;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -15,9 +16,12 @@ import java.util.List;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final SaleItemRepository saleItemRepository;
 
-    public ProductService(ProductRepository productRepository) {
+
+    public ProductService(ProductRepository productRepository, SaleItemRepository saleItemRepository) {
         this.productRepository = productRepository;
+        this.saleItemRepository = saleItemRepository;
     }
 
     public List<Product> findAll() {
@@ -34,7 +38,15 @@ public class ProductService {
     }
 
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        try {
+            if(saleItemRepository.existsByProductId(id)){
+                throw new RuntimeException("Cannot delete this product because it has sales records!");
+            }
+            productRepository.deleteById(id);
+            productRepository.flush();
+        } catch (Exception e) {
+            throw new RuntimeException("Cannot delete this product because it has sales records!");
+        }
     }
 
     public List<Product> findByName(String name) {
@@ -44,4 +56,9 @@ public class ProductService {
     public List<Product> findLowStock() {
         return productRepository.findByStockQuantityLessThanEqual(5);
     }
+
+    public Product findByBarCode(String barCode) {
+        return productRepository.findByBarcode(barCode);
+    }
+
 }
